Added new Shooter Control.vi
It has the new, revised joystick values now.  It includes all the
functions of the shooter arm, the shooter rotation, the shooter feeder,
and the shooter itself.  I incorporated it into Teleop and disabled the
old shooter control code.  I am not sure what values should be used for
the 4 levels of the shooter arm however.  Any advice on this?